qtbase: Made configuration more generic
Introduced linuxfb instead of glesv2/egl because linuxfb is more
generic. Please notice that some of the qt modules does not compile
without opengl support so you will need to overlay this qtbase Buildfile
with one containing proper opengl for your particular board. For example
qtdeclarative does not compile without the quick (2) module which is
only provided by qtbase if opengl is enabled.